requirements.txt
# Core
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
pyyaml==6.0.2
tenacity==9.0.0
httpx==0.27.2
rich==13.9.2
structlog==24.4.0

# LLM providers — all OpenAI-compatible, we use openai SDK as the common client
openai==1.54.0          # works for OpenAI + Groq + OpenRouter
anthropic==0.39.0       # native Anthropic path

# arXiv ingest + PDF
arxiv==2.1.3
pypdf==5.1.0

# Embeddings + rerankers (local, CPU-friendly)
sentence-transformers==3.2.1
torch==2.4.1            # installed from the CPU index by the bootstrap script

# Vector DB
qdrant-client==1.12.1

# BM25 hybrid retrieval
rank-bm25==0.2.2

# RAG eval
ragas==0.2.6
datasets==3.0.1

# Tracing
langfuse==2.55.0

# Serving / UI
fastapi==0.115.0
uvicorn[standard]==0.31.0
streamlit==1.39.0
plotly==5.24.1

# Dev
pytest==8.3.3
pytest-cov==5.0.0
pytest-mock==3.14.0
pytest-asyncio==0.24.0
black==24.10.0
ruff==0.7.1
mypy==1.12.1
pre-commit==4.0.1
